# perf: use V8 Deno builtins