--- a/core/runtime/jsruntime.rs
+++ b/core/runtime/jsruntime.rs
@@ -341,6 +341,7 @@ fn v8_init(
     " --no-validate-asm",
     " --turbo_fast_api_calls",
     " --harmony-change-array-by-copy",
+    " --expose_deno_builtins",
   );
   let predictable_flags = "--predictable --random-seed=42";
   let expose_natives_flags = "--expose_gc --allow_natives_syntax";