Upto 2x faster encode/decode

```
divy@mini ~/g/deno (main)> deno run -A bench.js
decode: Hello 😃
cpu: unknown
runtime: deno 1.36.4 (aarch64-apple-darwin)

benchmark      time (avg)             (min … max)    
--------------------------------------------------
            122.27 ns/iter  (122.02 ns … 215.82 ns)  

divy@mini ~/g/deno (main)> target/release/deno run -A bench.js
decode: Hello 😃
cpu: unknown
runtime: deno 1.36.4 (aarch64-apple-darwin)

benchmark      time (avg)             (min … max)       
---------------------------------------------------
            64.02 ns/iter  (62.81 ns … 139.55 ns) 
```